refactor: extract default server components
